Client Renamed - DW
In process of renaming the client project so we can overwrite the original base project that was contained inside. This commit will push the project folder under the name SamClient_Demon and remove the original SamClient folder from the repo. We will then rename SamClient_Demon to SamClient then commit again.