RP Changed - DW
Changed RP system to URP from HDRP for the server. This is mainly for in the case that an Apple laptop is required for usage on the day. Previous tests with HDRP resulted in blown out color values due to the way the shaders were being processed. This will also thankfully lower batches per frame and allow for smaller package size when shipping out.